+++ src/main/java/com/jcabi/log/MulticolorLayout.java
@@ -127,7 +127,7 @@ public final class MulticolorLayout extends EnhancedPatternLayout {
      */
     @SuppressWarnings("PMD.UseConcurrentHashMap")
     public void setColors(final String cols) {
-        final Map<String, String> parsed = new ParseInformation(
+        final Map<String, String> parsed = new ParseableInformation(
             cols
         ).parse();
         for (final Entry<String, String> entry : parsed.entrySet()) {
@@ -151,7 +151,7 @@ public final class MulticolorLayout extends EnhancedPatternLayout {
      */
     @SuppressWarnings("PMD.UseConcurrentHashMap")
     public void setLevels(final String lev) {
-        final Map<String, String> parsed = new ParseLevelInformation(
+        final Map<String, String> parsed = new ParseableLevelInformation(
             lev
         ).parse();
         for (final Entry<String, String> entry : parsed.entrySet()) {
